Integrate Google.Events into Google.Functions.Framework
Note that the framework does *not* depend on Google.Events.Protobuf or Google.Events.SystemTextJson - but our test projects do.
Until we publish Google.Events etc to NuGet, we can't really change the templates and examples to reflect this very easily. That can happen later - but it does mean the examples won't build locally right now.